Add batch file that builds and executes all GPR tests. Removed bad format test to fix gpr_string_test. Also took care of some of the MS compiler warnings when compiling gpr_test_util. 	Change on 2014/12/11 by jtattermusch <jtattermusch@google.com> ------------- Created by MOE: http://code.google.com/p/moe-java MOE_MIGRATED_REVID=81900639

+++ test/core/end2end/cq_verifier.c
@@ -133,7 +133,7 @@ static void verify_and_destroy_metadata(metadata *md, grpc_metadata *elems,
   gpr_free(md);
 }
 
-static gpr_slice merge_slices(gpr_slice *slices, int nslices) {
+static gpr_slice merge_slices(gpr_slice *slices, size_t nslices) {
   size_t i;
   size_t len = 0;
   gpr_uint8 *cursor;
